refactor: service 내 함수들 각 handlers로 분리
