+++ src/js/handlers/videoSave.js
@@ -1,7 +1,22 @@
 import { EXCEED_SAVED_VIDEO_COUNT_MSG } from '../constants/alertMessage.js';
 import { MAX_SAVED_VIDEO_COUNT } from '../constants/classroom.js';
-import { saveVideo } from '../service.js';
 import videoInfos from '../states/videoInfos.js';
+
+function createVideoInfo(videoDataset) {
+  const { videoId, title, channelId, channelTitle, publishTime } = videoDataset;
+
+  return {
+    id: { videoId },
+    snippet: { title, channelId, channelTitle, publishTime },
+    isWatched: false,
+  };
+}
+
+function saveVideo($video) {
+  const videoInfo = createVideoInfo($video.dataset);
+
+  videoInfos.add(videoInfo);
+}
 
 function handleVideoSave({ target }) {
   if (!target.classList.contains('js-save-button')) return;
